requirements.txt
rasterio
numpy
Flask
Pillow
matplotlib
openai
shapely
fpdf2
matplotlib
